fix bug confirming model

+++ train/train_confirming.py
@@ -90,7 +90,7 @@ def train(model, device, train_loader, criterion, optimizer, scheduler, epoch, i
             
             label_pred = decoder(output)
 
-            train_precision = precision_score(np.array(label_pred, dtype=np.float32), np.array(labels.tolist(), dtype=np.float32), average='micro')
+            train_precision = precision_score(np.array(label_pred), np.array(labels.tolist()), average='micro')
             train_precision_average.append(train_precision)
 
             optimizer.step()
@@ -127,7 +127,7 @@ def test(model, device, test_loader, criterion, iter_meter, experiment, filename
 
                 label_pred[0] = round(label_pred[0], 10)
 
-                test_precision = precision_score(np.array(label_pred, dtype=np.float32), np.array(labels.tolist(), dtype=np.float32), average='micro')
+                test_precision = precision_score(np.array(label_pred), np.array(labels.tolist()), average='micro')
                 test_precision_average.append(test_precision)
 
     experiment.log_metric('test_loss', test_loss, step=iter_meter.get())
